Got server-side rendering to work
