+++ packages/codemirror-6/.gitignore
@@ -1,1 +1,2 @@
 dist
+*.swp
