+++ packages/experiments/demo/demo.ts
@@ -1,42 +1,3 @@
-import {
-  history,
-  EditorState,
-  EditorSelection
-  EditorView,
-  keymap,
-  gutter,
-  baseKeymap,
-  legacyMode,
-  matchBrackets,
-  javascript,
-  specialChars,
-  multipleSelections
-} from 'codemirror-6';
-
-import { historyKeymap } from './historyKeymap';
-import { indentationKeymap } from './indentationKeymap';
-import { experimentPlugin } from './experimentPlugin';
-import { ot } from 'codemirror-ot';
-
-let mode = legacyMode(javascript({indentUnit: 2}, {}))
-let state = EditorState.create({doc: `"use strict";
-const {readFile} = require("fs");
-
-readFile("package.json", "utf8", (err, data) => {
-  console.log(data);
-});`, plugins: [
-  gutter(),
-  history(),
-  specialChars({}),
-  multipleSelections(),
-  mode,
-  matchBrackets({decorationsPlugin: mode}),
-  keymap(historyKeymap()),
-  keymap(indentationKeymap(mode)),
-  keymap(baseKeymap),
-  experimentPlugin,
-  ot
-]})
-
-let view = (window as any).view = new EditorView(state)
+import { createView } from './demoView';
+const view = createView();
 document.querySelector("#editor").appendChild(view.dom)
