+++ packages/experiments/demo/indentationKeymap.ts
@@ -1,3 +1,7 @@
+import {
+  EditorView,
+  EditorSelection
+} from 'codemirror-6';
 export const indentationKeymap = mode => {
   function crudeInsertNewlineAndIndent({state, dispatch}: EditorView): boolean {
     let indentation = (mode as any).indentation(state, state.selection.primary.from)
