+++ packages/experiments/demo/server.ts
@@ -1,12 +1,29 @@
 import * as express  from 'express';
 import { Application, Router, Request, Response } from 'express';
 import * as jsdom from 'jsdom';
+import { createView } from './demoView';
 
 const { JSDOM } = jsdom;
-const html = '<!DOCTYPE html><p>Hello JSDOM</p>'
+const html = '<!DOCTYPE html><div id="editor"></div>'
 
 const document = new JSDOM(html).window.document;
-//global.document = new JSDOM(html).window.document;
+//const documentAny: any = documentJSDOM;
+//const document: HTMLElement = documentAny;
+document.getSelection = () => ({}) as Selection;
+const globalAny:any = global;
+globalAny.document = document;
+globalAny.navigator = {};
+globalAny.window = {
+  addEventListener: () => {}
+};
+globalAny.MutationObserver = () => ({
+  observe: () => {},
+  takeRecords: () => {},
+  disconnect: () => {}
+});
+globalAny.requestAnimationFrame = () => {};
+const view = createView();
+document.querySelector("#editor").appendChild(view.dom)
 
 const router: Router = Router();
