+++ packages/experiments/tsconfig.json
@@ -4,7 +4,7 @@
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "sourceMap": true,
-    "strict": true,
+    "strict": false,
     "target": "es5",
     "newLine": "lf"
   },
